Add files via upload
audio.js provides the ability to add new audio tracks via URL and provides direct audio control and modification of the playback speed. 

mainMenu.js is a script that contains the current placeholder menus for the game, the pause screen and the main menu that are interacted with using the arrow keys and ENTER 

documentation and more specific descriptions are located at the top of the script in comments.