Added test levels and rotated assets + collisions
Overall just made some test levels. FYI, level 3 completion will lead to
white screen, Kyle that has to do with two buttons script expecting the
level to be level 1. Level 4 is supposed to be decently long so that is
why it is unfinished.
Kyle: I added in checks in twoButtons.js to check if you are in level 1
OR level 3. Overall we need to change things a bit to incorporate other
levels.
Jarrett: I made rotated versions of assets and renamed them in the
scripts they were called in. We need to figure out how to have door
objects work in levels other than the 1st.