added level 1 with changes + attempted to fix oneButton
added level 1 with collisons and everything but does not transition
properly. Old level 1 (two buttons) is now level 11 and the assets exist
the same as they were. This also meant the the game objects for buttons
and doors and stuff are now numbered according to the level they belong
to.
Kyle: The one button code has the player automatically transitioning to
a new level upon press. It also for some reason does not think openDoor
is a function and I had to paste the function code in directly in the
collision event for it work.
Jarrett: when trying to do level 11 it throws an error saying
parent.addComponent doesen't exist. I don't know why it is doing this
but just wanted to let you know in case it complicates any level after.